The '$' in the class name of Digester$EnvironmentPropertySource is not really shell frendly for unix environments. The class should not a inner class.
