+++ webapps/docs/config/systemprops.xml
@@ -49,7 +49,7 @@
       <p>Property replacement from the specified property source on the JVM
          system properties can also be done using the
          <code>REPLACE_SYSTEM_PROPERTIES</code> system property.</p>
-      <p><code>org.apache.tomcat.util.digester.Digester$EnvironmentPropertySource</code>
+      <p><code>org.apache.tomcat.util.digester.EnvironmentPropertySource</code>
          can be used to replace parameters from the process' environment
          variables, e.g. injected ConfigMaps or Secret objects in container
          based systems like OpenShift or Kubernetes.</p>
